+++ src/main/java/com/jiazhe/youxiang/server/dto/chargeoff/ChargeOffInfoDTO.java
@@ -51,7 +51,7 @@ public class ChargeOffInfoDTO {
     /**
      * 核销总积分
      */
-    private Integer totalPoint;
+    private BigDecimal totalPoint;
 
     /**
      * 客户姓名
@@ -141,11 +141,11 @@ public class ChargeOffInfoDTO {
         this.pointList = pointList;
     }
 
-    public Integer getTotalPoint() {
+    public BigDecimal getTotalPoint() {
         return totalPoint;
     }
 
-    public void setTotalPoint(Integer totalPoint) {
+    public void setTotalPoint(BigDecimal totalPoint) {
         this.totalPoint = totalPoint;
     }
