+++ src/main/java/com/jiazhe/youxiang/server/vo/req/chargeoff/ChargeOffAddReq.java
@@ -33,7 +33,7 @@ public class ChargeOffAddReq extends BaseVO {
     private List<String> keytList;
 
     @ApiModelProperty("核销总积分")
-    private Integer totalPoint;
+    private BigDecimal totalPoint;
 
     @ApiModelProperty("客户姓名")
     private String customerName;
@@ -82,11 +82,11 @@ public class ChargeOffAddReq extends BaseVO {
         this.keytList = keytList;
     }
 
-    public Integer getTotalPoint() {
+    public BigDecimal getTotalPoint() {
         return totalPoint;
     }
 
-    public void setTotalPoint(Integer totalPoint) {
+    public void setTotalPoint(BigDecimal totalPoint) {
         this.totalPoint = totalPoint;
     }
