+++ src/main/java/com/jiazhe/youxiang/server/vo/resp/chargeoff/ChargeOffInfoResp.java
@@ -39,7 +39,7 @@ public class ChargeOffInfoResp extends BaseVO {
     private List<ChargeOffPointResp> pointList;
 
     @ApiModelProperty("核销总积分")
-    private Integer totalPoint;
+    private BigDecimal totalPoint;
 
     @ApiModelProperty("客户姓名")
     private String customerName;
@@ -113,11 +113,11 @@ public class ChargeOffInfoResp extends BaseVO {
         this.pointList = pointList;
     }
 
-    public Integer getTotalPoint() {
+    public BigDecimal getTotalPoint() {
         return totalPoint;
     }
 
-    public void setTotalPoint(Integer totalPoint) {
+    public void setTotalPoint(BigDecimal totalPoint) {
         this.totalPoint = totalPoint;
     }
